Refactor: 인터페이스와 구현체 분리
Swagger 에서 사용하는 Annotation을 인터페이스쪽으로 분리하고, 각 컴포넌트의 역할을 인터페이스에 명시한다.
